+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [Unreleased]
 
+- Mark peripheral with `non_exhaustive`
 - Force using rust edition 2021 in CI
 - Added lifetime ellision for `FieldWriter` where the explicit lifetimes are not necessary, which
   fixes the `clippy::needless_lifetimes` warning on rustc 1.84